Create dynamic.test_models module and clean up imports
