+++ webapp/apps/taxbrain/tests/test_models.py
@@ -1,23 +1,10 @@
 from django.test import TestCase
-from django.test import Client
-import json
-import os
 import pytest
-import numpy as np
-from datetime import datetime
 
-from ..models import (JSONReformTaxCalculator,
-                      OutputUrl)
+from ..models import JSONReformTaxCalculator
 from ...test_assets.utils import get_taxbrain_model
 from ...test_assets.test_models import TaxBrainTableResults
-from ..forms import PersonalExemptionForm
 
-from ...dynamic.models import DynamicBehaviorOutputUrl
-from ...dynamic.forms import DynamicBehavioralInputsModelForm
-
-START_YEAR = 2016
-
-CURDIR = os.path.abspath(os.path.dirname(__file__))
 
 class TaxBrainJSONReformModelTest(TestCase):
     """Test taxbrain JSONReformTaxCalculator."""
@@ -35,19 +22,10 @@ class TaxBrainJSONReformModelTest(TestCase):
         )
 
 @pytest.mark.usefixtures("test_coverage_fields")
-class TaxBrainResultsTest(TaxBrainTableResults, TestCase):
+class TaxBrainStaticResultsTest(TaxBrainTableResults, TestCase):
 
     def test_static_tc_lt_0130(self):
         self.tc_lt_0130()
 
     def test_static_tc_gt_0130(self):
         self.tc_gt_0130()
-
-
-    def test_dynamic_tc_lt_0130(self):
-        self.tc_lt_0130(Form=DynamicBehavioralInputsModelForm,
-                        UrlModel=DynamicBehaviorOutputUrl)
-
-    def test_dynamic_tc_gt_0130(self):
-        self.tc_gt_0130(Form=DynamicBehavioralInputsModelForm,
-                        UrlModel=DynamicBehaviorOutputUrl)
